+++ src/components/metric/metric.module.ts
@@ -21,7 +21,7 @@ import { NgModule} from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { MetricComponent } from './metric.component';
 import { TranslateModule } from '@ngx-translate/core';
-import { FormatNumberModule } from '../../pipes/pipes.module';
+import { FormatNumberModule } from '../../pipes/format-number/format-number.module';
 
 
 @NgModule({
