+++ src/components/powerbars/powerbars.module.ts
@@ -23,7 +23,7 @@ import { PowerbarsComponent } from './powerbars.component';
 import { MatCardModule, MatIconModule, MatTooltipModule, MatInputModule } from '@angular/material';
 import { TranslateModule } from '@ngx-translate/core';
 import { ColorGeneratorModule } from '../../services/color.generator.module';
-import { FormatNumberModule } from '../../pipes/pipes.module';
+import { FormatNumberModule } from '../../pipes/format-number/format-number.module';
 
 @NgModule({
   imports: [
