+++ src/pipes/format-number/format-number.pipe.ts
@@ -16,17 +16,19 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import { FormatNumberPipe } from './fomat-number.pipe';
-import { NgModule } from '@angular/core';
 
-@NgModule({
-  imports: [
-  ],
-  declarations: [
-    FormatNumberPipe
-  ],
-  exports: [
-    FormatNumberPipe
-  ]
+import { Pipe, PipeTransform } from '@angular/core';
+import { formatNumber } from '../../components/componentsUtils';
+import { TranslateService } from '@ngx-translate/core';
+
+@Pipe({
+  name: 'formatNumber'
 })
-export class FormatNumberModule {}
+export class FormatNumberPipe implements PipeTransform {
+
+  constructor(private translate: TranslateService) {}
+  public transform(x, formatChar: string = ' '): any {
+    return formatNumber(x, this.translate.instant(formatChar));
+  }
+
+}
